Create Mother Vertex
Here the task is to find Mother Vertex, a mother vertex in a graph G is a vertex v such that all other vertices in G can be reached by a path from v. Here first DFS algorithm is used and inside the first loop of 'findMother' method, 'v' stores last finished vertex. If there exist mother vertex (or vertices), then one of the mother vertices is the last finished vertex in DFS. Hence DFS method is called again to check if 'v' is mother vertex or not and after that 'visited' array is used for checking the same.