Create Strongly Connected Components Kosarajus Algorithm
A directed graph is called strongly connected if there is a path in each direction between each pair of vertices of the graph. That is, a path exists from the first vertex in the pair to the second, and another path exists from the second vertex to the first. Here, instead of 'n', 'n-1' is used because from the main function of GeeksForGeeks, 'n+1' was passed as argument to 'kosaraju' function. 

Here, first 'fillStack' function is called which is specialized version of DFS (It also pushes vertices in Stack but in reverse order). Then transpose of Graph is calculated and DFS is called for each vertex in Stack.