Create Sum Of Bit Differences
Here the task, given an integer array of n integers, we have tp find sum of bit differences in all pairs that can be formed from array elements. Bit difference of a pair (x, y) is count of different bits at same positions in binary representations of x and y. This solution is only for numbers that can be represented in maximum of '4' bits.