Create Inversion Of Array
Here the task is to find the inversion count of the array.
Inversion Count : For an array, inversion count indicates how far (or close) the array is from being sorted. If array is already sorted then inversion count is 0. If array is sorted in reverse order that inversion count is the maximum. 
Formally, two elements a[i] and a[j] form an inversion if a[i] > a[j] and i < j.

In this program, merge sort is used and the expected time complexity is O(nlogn).