Create Celebrity Problem
Here Stack is used and the expected time complexity is O(3(N-1)) because check() method is called 3(N-1) number of times. First all the person are pushed on to stack. Then every time 2 persons are popped, it is checked which of them can be a celebrity by using check() method, person which can be a celebrity is again pushed on to stack. This process is repeated till there is only one person left in the stack. Then that person is popped and using a loop, it is checked that (s)he is a celebrity or not.