Create Partition Array Into K Equal Sum Subsets
Here 'used' is a boolean array which is used to track which elements of the array are already used in forming subets. Another important thing to note is 'visited' 2D boolean array. It is used to keep track of 'current sum' which will not lead us to 'sum / k'. For convenience I have marked it to true if 'current sum' can't lead us to target sum i.e. 'sum / k'. This solution is blend of both 'Dynamic Programming' and 'Backtracking'.