Create Largest Square Formed In A Matrix
Given a binary matrix, find out the maximum size square sub-matrix with all 1s. Here, a boolean variable 'flag' is used to check whether there is at least one '1' in matrix in case if there is either a single row and multiple columns or a single column and multiple rows, if so the output will be '1'. If there are multiple rows and multiple columns then we just have to print value of 'max'.