Create Minimum Number Of Deletions To Make A String Palindrome
Task is to write a program to remove or delete minimum number of characters from the string so that the resultant string is palindrome. The Solution is to find the length of Longest Palindromic Subsequence and subtract it from length of the original string, the resultant length will be the answer.