Create Find Number Of Times A String Occurs As A Subsequence
Given two strings S1 and S2, we have to find the number of times the second string occurs in the first string, whether continuous or discontinuous. Here, second string S2 is traversed in reverse order and its each character is compared with each character of the main string S1 from left to right.