Create Special Keyboard
Here the by dry run, I found that for (n <= 6), the answer is 'n'. If (n > 6) then, the answer can be maximum of (1 + count(n - 1)) which is adding 'A' to the answer of count(n - 1) or the answer can be reached by repeatedly pasting after copying at some previous maximum. So we search for the previous maximum starting from (i = 3) to (n - 1) since copy, select and paste will take three steps so we go back 3 times at least.